refactor: :recycle: shift contact creation into model
The internal state of several components was not correctly updating/resetting when some dialogs would close. this is a massive refactor to fix some of those underlying issues as well as move contact creation/updating out of the dialog so that we can programatically create a contact without user interaction, specifically useful for creating a crowdnode or maya contact.